support for different machines

+++ .config/polybar/launch.sh
@@ -1,18 +1,36 @@
-#!/usr/bin/env sh
+#!/usr/bin/env bash
 
-# Add to i3 startup
+function set_env_vars() {
+    export NETWORK_INTERFACE=$(ip route | grep '^default' | awk '{print $5}' | head -n1)
+    export BATTERY=$(ls /sys/class/power_supply/ | grep 'BAT[[:digit:]]')
+    export ADAPTER=$(ls /sys/class/power_supply/ | grep 'ADP[[:digit:]]')
+}
 
-# Terminate already running bar instances
-pkill polybar
+function stop() {
+    if pgrep polybar ; then
+        # terminate all polybar instances
+        pkill polybar
+        # wait for process to shut down
+        while pgrep -u $UID -x polybar > /dev/null ; do 
+            sleep 0.1 
+        done
+    fi
+}
 
-# Wait until the processes have been shut down
-while pgrep -u $UID -x polybar >/dev/null; do sleep 1; done
+function start() {
+    if type "xrandr" ; then
+        # launch bar on each monitor
+        for m in $(xrandr --query | grep " connected" | cut -d" " -f1) ; do
+            MONITOR=$m polybar -c ~/.config/polybar/config.ini main &
+        done
+    else
+        # launch one bar instance
+        polybar -c ~/.config/polybar/config.ini main &
+    fi
+}
 
-# Launch bars on each monitor
-if type "xrandr"; then
-  for m in $(xrandr --query | grep " connected" | cut -d" " -f1); do
-    MONITOR=$m polybar -c ~/.config/polybar/config.ini main &
-  done
-else
-  polybar -c ~/.config/polybar/config.ini main &
-fi
+set_env_vars
+
+stop
+
+start
